+++ src/base/widget.c
@@ -1165,8 +1165,6 @@ ret_t widget_set_prop(widget_t* widget, const char* name, const value_t* v) {
     widget->enable = value_bool(v);
   } else if (tk_str_eq(name, WIDGET_PROP_NAME)) {
     widget_set_name(widget, value_str(v));
-  } else if (tk_str_eq(name, WIDGET_PROP_TEXT)) {
-    wstr_from_value(&(widget->text), v);
   } else if (tk_str_eq(name, WIDGET_PROP_TR_TEXT)) {
     widget_set_tr_text(widget, value_str(v));
   } else if (tk_str_eq(name, WIDGET_PROP_ANIMATION)) {
@@ -1189,6 +1187,9 @@ ret_t widget_set_prop(widget_t* widget, const char* name, const value_t* v) {
   if (ret == RET_NOT_FOUND) {
     if (tk_str_eq(name, WIDGET_PROP_FOCUS)) {
       widget_set_focused(widget, value_bool(v));
+      ret = RET_OK;
+    } else if (tk_str_eq(name, WIDGET_PROP_TEXT)) {
+      wstr_from_value(&(widget->text), v);
       ret = RET_OK;
     } else if (tk_str_start_with(name, "style:")) {
       return widget_set_style(widget, name + 6, v);