+++ tests/edit_test.cc
@@ -7,6 +7,7 @@
 #include "widgets/edit.h"
 #include "widgets/window.h"
 #include "widgets/group_box.h"
+#include "common.h"
 
 TEST(Edit, int) {
   value_t v1;
@@ -401,4 +402,34 @@ TEST(Edit, is_valid_chr_4_custom) {
   ASSERT_EQ(edit_input_char(e, L'a'), RET_OK);
 
   widget_destroy(e);
+}
+
+TEST(Edit, events) {
+  string event_log;
+  widget_t* b = edit_create(NULL, 10, 20, 30, 40);
+
+  widget_on(b, EVT_VALUE_CHANGING, widget_log_events, &event_log);
+  widget_on(b, EVT_VALUE_CHANGED, widget_log_events, &event_log);
+
+  event_log = "";
+  widget_set_text(b, L"a");
+  ASSERT_EQ(event_log, "value_changed");
+  ASSERT_EQ(wcscmp(b->text.str, L"a"), 0);
+
+  event_log = "";
+  widget_set_text(b, L"a");
+  ASSERT_EQ(event_log, "");
+  ASSERT_EQ(wcscmp(b->text.str, L"a"), 0);
+
+  event_log = "";
+  widget_set_text(b, L"1");
+  ASSERT_EQ(event_log, "value_changed");
+  ASSERT_EQ(wcscmp(b->text.str, L"1"), 0);
+
+  event_log = "";
+  edit_input_char(b, 'a');
+  ASSERT_EQ(event_log, "value_changing");
+  ASSERT_EQ(wcscmp(b->text.str, L"1a"), 0);
+
+  widget_destroy(b);
 }