+++ tests/widget_test.cc
@@ -16,6 +16,7 @@
 #include "gtest/gtest.h"
 
 using std::string;
+#include "common.h"
 
 TEST(Widget, basic1) {
   widget_t* w = window_create(NULL, 0, 0, 400, 300);
@@ -219,125 +220,54 @@ TEST(Widget, children) {
   widget_destroy(w);
 }
 
-static string s_event_log;
-
-static ret_t on_button_events(void* ctx, event_t* e) {
-  widget_t* widget = WIDGET(e->target);
-  assert(widget->can_not_destroy > 0);
-
-  (void)ctx;
-  switch (e->type) {
-    case EVT_MOVE: {
-      s_event_log += "move ";
-      break;
-    }
-    case EVT_WILL_MOVE: {
-      s_event_log += "will_move ";
-      break;
-    }
-    case EVT_RESIZE: {
-      s_event_log += "resize ";
-      break;
-    }
-    case EVT_WILL_RESIZE: {
-      s_event_log += "will_resize ";
-      break;
-    }
-    case EVT_MOVE_RESIZE: {
-      s_event_log += "move_resize ";
-      break;
-    }
-    case EVT_WILL_MOVE_RESIZE: {
-      s_event_log += "will_move_resize ";
-      break;
-    }
-    case EVT_PROP_WILL_CHANGE: {
-      prop_change_event_t* evt = (prop_change_event_t*)e;
-
-      s_event_log += "prop_will_change ";
-      s_event_log += evt->name;
-      s_event_log += value_str(evt->value);
-      s_event_log += " ";
-      break;
-    }
-    case EVT_PROP_CHANGED: {
-      prop_change_event_t* evt = (prop_change_event_t*)e;
-
-      s_event_log += "prop_changed ";
-      s_event_log += evt->name;
-      s_event_log += value_str(evt->value);
-      s_event_log += " ";
-      break;
-    }
-    case EVT_POINTER_MOVE: {
-      s_event_log += "EVT_POINTER_MOVE";
-      break;
-    }
-    case EVT_POINTER_DOWN: {
-      s_event_log += "EVT_POINTER_DOWN";
-      break;
-    }
-    case EVT_POINTER_UP: {
-      s_event_log += "EVT_POINTER_UP";
-      break;
-    }
-    case EVT_KEY_DOWN: {
-      s_event_log += "EVT_KEY_DOWN";
-      break;
-    }
-    case EVT_KEY_UP: {
-      s_event_log += "EVT_KEY_UP";
-      break;
-    }
-  }
-  return RET_OK;
-}
-
 TEST(Widget, move) {
+  string event_log;
   widget_t* w = button_create(NULL, 0, 0, 0, 0);
 
-  s_event_log = "";
-  widget_on(w, EVT_MOVE, on_button_events, NULL);
-  widget_on(w, EVT_WILL_MOVE, on_button_events, NULL);
+  event_log = "";
+  widget_on(w, EVT_MOVE, widget_log_events, &event_log);
+  widget_on(w, EVT_WILL_MOVE, widget_log_events, &event_log);
 
   widget_move(w, 100, 200);
   ASSERT_EQ(w->x, 100);
   ASSERT_EQ(w->y, 200);
-  ASSERT_EQ(s_event_log, string("will_move move "));
+  ASSERT_EQ(event_log, string("will_move move "));
   ASSERT_EQ(w->can_not_destroy, 0);
 
   widget_destroy(w);
 }
 
 TEST(Widget, resize) {
+  string event_log;
   widget_t* w = button_create(NULL, 0, 0, 0, 0);
 
-  s_event_log = "";
-  widget_on(w, EVT_RESIZE, on_button_events, NULL);
-  widget_on(w, EVT_WILL_RESIZE, on_button_events, NULL);
+  event_log = "";
+  widget_on(w, EVT_RESIZE, widget_log_events, &event_log);
+  widget_on(w, EVT_WILL_RESIZE, widget_log_events, &event_log);
 
   widget_resize(w, 100, 200);
   ASSERT_EQ(w->w, 100);
   ASSERT_EQ(w->h, 200);
-  ASSERT_EQ(s_event_log, string("will_resize resize "));
+  ASSERT_EQ(event_log, string("will_resize resize "));
   ASSERT_EQ(w->can_not_destroy, 0);
 
   widget_destroy(w);
 }
 
 TEST(Widget, move_resize) {
+  string event_log;
   widget_t* w = button_create(NULL, 0, 0, 0, 0);
 
-  s_event_log = "";
-  widget_on(w, EVT_MOVE_RESIZE, on_button_events, NULL);
-  widget_on(w, EVT_WILL_MOVE_RESIZE, on_button_events, NULL);
+  event_log = "";
+  widget_on(w, EVT_MOVE_RESIZE, widget_log_events, &event_log);
+  widget_on(w, EVT_WILL_MOVE_RESIZE, widget_log_events, &event_log);
 
   widget_move_resize(w, 100, 200, 300, 400);
   ASSERT_EQ(w->x, 100);
   ASSERT_EQ(w->y, 200);
   ASSERT_EQ(w->w, 300);
   ASSERT_EQ(w->h, 400);
-  ASSERT_EQ(s_event_log, string("will_move_resize move_resize "));
+  ASSERT_EQ(event_log, string("will_move_resize move_resize "));
   ASSERT_EQ(w->can_not_destroy, 0);
 
   widget_destroy(w);
@@ -345,15 +275,16 @@ TEST(Widget, move_resize) {
 
 TEST(Widget, prop) {
   value_t v;
+  string event_log;
   widget_t* w = button_create(NULL, 0, 0, 0, 0);
 
-  widget_on(w, EVT_PROP_WILL_CHANGE, on_button_events, NULL);
-  widget_on(w, EVT_PROP_CHANGED, on_button_events, NULL);
+  widget_on(w, EVT_PROP_WILL_CHANGE, widget_log_events, &event_log);
+  widget_on(w, EVT_PROP_CHANGED, widget_log_events, &event_log);
 
-  s_event_log = "";
+  event_log = "";
   value_set_str(&v, "123");
   ASSERT_EQ(widget_set_prop(w, "name", &v), RET_OK);
-  ASSERT_EQ(s_event_log, string("prop_will_change name123 prop_changed name123 "));
+  ASSERT_EQ(event_log, string("prop_will_change name123 prop_changed name123 "));
 
   widget_destroy(w);
 }
@@ -664,25 +595,26 @@ TEST(Widget, get_window) {
 }
 
 static ret_t on_button_events_stop(void* ctx, event_t* e) {
+  string& event_log = *(string*)ctx;
   switch (e->type) {
     case EVT_POINTER_MOVE_BEFORE_CHILDREN: {
-      s_event_log += "EVT_POINTER_MOVE_BEFORE_CHILDREN";
+      event_log += "EVT_POINTER_MOVE_BEFORE_CHILDREN";
       break;
     }
     case EVT_POINTER_DOWN_BEFORE_CHILDREN: {
-      s_event_log += "EVT_POINTER_DOWN_BEFORE_CHILDREN";
+      event_log += "EVT_POINTER_DOWN_BEFORE_CHILDREN";
       break;
     }
     case EVT_POINTER_UP_BEFORE_CHILDREN: {
-      s_event_log += "EVT_POINTER_UP_BEFORE_CHILDREN";
+      event_log += "EVT_POINTER_UP_BEFORE_CHILDREN";
       break;
     }
     case EVT_KEY_DOWN_BEFORE_CHILDREN: {
-      s_event_log += "EVT_KEY_DOWN_BEFORE_CHILDREN";
+      event_log += "EVT_KEY_DOWN_BEFORE_CHILDREN";
       break;
     }
     case EVT_KEY_UP_BEFORE_CHILDREN: {
-      s_event_log += "EVT_KEY_UP_BEFORE_CHILDREN";
+      event_log += "EVT_KEY_UP_BEFORE_CHILDREN";
       break;
     }
     default:
@@ -693,76 +625,81 @@ static ret_t on_button_events_stop(void* ctx, event_t* e) {
 }
 
 TEST(Widget, EVT_POINTER_DOWN_BEFORE_CHILDREN) {
+  string event_log;
   pointer_event_t evt;
   widget_t* w = button_create(NULL, 0, 0, 0, 0);
 
-  s_event_log = "";
+  event_log = "";
   pointer_event_init(&evt, EVT_POINTER_DOWN, w, 1, 2);
-  widget_on(w, EVT_POINTER_DOWN_BEFORE_CHILDREN, on_button_events_stop, NULL);
-  widget_on(w, EVT_POINTER_DOWN, on_button_events, NULL);
+  widget_on(w, EVT_POINTER_DOWN_BEFORE_CHILDREN, on_button_events_stop, &event_log);
+  widget_on(w, EVT_POINTER_DOWN, widget_log_events, &event_log);
   widget_on_pointer_down(w, &evt);
 
-  ASSERT_EQ(s_event_log, string("EVT_POINTER_DOWN_BEFORE_CHILDREN"));
+  ASSERT_EQ(event_log, string("EVT_POINTER_DOWN_BEFORE_CHILDREN"));
 
   widget_destroy(w);
 }
 
 TEST(Widget, EVT_POINTER_UP_BEFORE_CHILDREN) {
+  string event_log;
   pointer_event_t evt;
   widget_t* w = button_create(NULL, 0, 0, 0, 0);
 
-  s_event_log = "";
+  event_log = "";
   pointer_event_init(&evt, EVT_POINTER_UP, w, 1, 2);
-  widget_on(w, EVT_POINTER_UP_BEFORE_CHILDREN, on_button_events_stop, NULL);
-  widget_on(w, EVT_POINTER_UP, on_button_events, NULL);
+  widget_on(w, EVT_POINTER_UP_BEFORE_CHILDREN, on_button_events_stop, &event_log);
+  widget_on(w, EVT_POINTER_UP, widget_log_events, &event_log);
   widget_on_pointer_up(w, &evt);
 
-  ASSERT_EQ(s_event_log, string("EVT_POINTER_UP_BEFORE_CHILDREN"));
+  ASSERT_EQ(event_log, string("EVT_POINTER_UP_BEFORE_CHILDREN"));
 
   widget_destroy(w);
 }
 
 TEST(Widget, EVT_POINTER_MOVE_BEFORE_CHILDREN) {
+  string event_log;
   pointer_event_t evt;
   widget_t* w = button_create(NULL, 0, 0, 0, 0);
 
-  s_event_log = "";
+  event_log = "";
   pointer_event_init(&evt, EVT_POINTER_MOVE, w, 1, 2);
-  widget_on(w, EVT_POINTER_MOVE_BEFORE_CHILDREN, on_button_events_stop, NULL);
-  widget_on(w, EVT_POINTER_MOVE, on_button_events, NULL);
+  widget_on(w, EVT_POINTER_MOVE_BEFORE_CHILDREN, on_button_events_stop, &event_log);
+  widget_on(w, EVT_POINTER_MOVE, widget_log_events, &event_log);
   widget_on_pointer_move(w, &evt);
 
-  ASSERT_EQ(s_event_log, string("EVT_POINTER_MOVE_BEFORE_CHILDREN"));
+  ASSERT_EQ(event_log, string("EVT_POINTER_MOVE_BEFORE_CHILDREN"));
 
   widget_destroy(w);
 }
 
 TEST(Widget, EVT_KEY_DOWN_BEFORE_CHILDREN) {
   key_event_t evt;
+  string event_log;
   widget_t* w = button_create(NULL, 0, 0, 0, 0);
 
-  s_event_log = "";
+  event_log = "";
   key_event_init(&evt, EVT_KEY_DOWN, w, 1);
-  widget_on(w, EVT_KEY_DOWN_BEFORE_CHILDREN, on_button_events_stop, NULL);
-  widget_on(w, EVT_KEY_DOWN, on_button_events, NULL);
+  widget_on(w, EVT_KEY_DOWN_BEFORE_CHILDREN, on_button_events_stop, &event_log);
+  widget_on(w, EVT_KEY_DOWN, widget_log_events, &event_log);
   widget_on_keydown(w, &evt);
 
-  ASSERT_EQ(s_event_log, string("EVT_KEY_DOWN_BEFORE_CHILDREN"));
+  ASSERT_EQ(event_log, string("EVT_KEY_DOWN_BEFORE_CHILDREN"));
 
   widget_destroy(w);
 }
 
 TEST(Widget, EVT_KEY_UP_BEFORE_CHILDREN) {
   key_event_t evt;
+  string event_log;
   widget_t* w = button_create(NULL, 0, 0, 0, 0);
 
-  s_event_log = "";
+  event_log = "";
   key_event_init(&evt, EVT_KEY_UP, w, 1);
-  widget_on(w, EVT_KEY_UP_BEFORE_CHILDREN, on_button_events_stop, NULL);
-  widget_on(w, EVT_KEY_UP, on_button_events, NULL);
+  widget_on(w, EVT_KEY_UP_BEFORE_CHILDREN, on_button_events_stop, &event_log);
+  widget_on(w, EVT_KEY_UP, widget_log_events, &event_log);
   widget_on_keyup(w, &evt);
 
-  ASSERT_EQ(s_event_log, string("EVT_KEY_UP_BEFORE_CHILDREN"));
+  ASSERT_EQ(event_log, string("EVT_KEY_UP_BEFORE_CHILDREN"));
 
   widget_destroy(w);
 }